Hooked off CKEditor changes, only highlight modified codeblocks
The CKEditor library BalloonEditor.create is now monkey patched to initialize
the created CKEditor instance Trilium creates with the marker conversion, and a
change postfixer in charge of highlighting. This is the simplest and most
efficient way of tracking CKEditors (and CKEditor changes), other than modifying
Trilium sources.

Note changes are tracked by that CKEDitor postfixer, which is easier although
less detailed than a CKEditor on change callback. Only codeblocks affected by
those changes are then re-highlighted.

Markers need to have an id unique to the editor (ie across all codeblocks of the
editor). This is currently just an autoincrementing sequence number that wraps
around some high value. Seven digit random numbers were found to be not unique
enough and cause collisions on notes with ~600 markers (such a note is not out
of the ordinary, since each line can be multiple markers, and id needs to be
unique be across codeblocks).

Removed hooks from the widget event handlers, the widget facilities are actually
not used, this could be converted to a #frontendStartup script.

Added revamped debug output functions.

Ignored highlighting if on plaintext type codeblocks, used auto on the rest.

Tests:
- splits
- changing note type
- changing readonly and back
- changing momentarily writeable, verified codeblocks are highlighted when that
  happens
- verified plaintext and non-plaintext (auto) codeblocks
- verified only modified codeblocks are rehighlighted